+++ library/src/types/baseTask.ts
@@ -1,0 +1,5 @@
+export interface BaseTask {
+  id: number | string;
+  startTime: Date;
+  endTime: Date;
+}
